Refactor project for task suit setup (Lecture 148)
The database fixture file was added in order to re-use setup and seed data for task test suit. The test script was also modified so that test suites for users and tasks run in series
