+++ .gitignore
@@ -62,3 +62,6 @@ typings/
 
 # Config files
 config/
+
+# .DS_Store
+.DS_Store
